HT36 Erase student id from job offer

+++ HireTalent/DAO's/StudentDAO.swift
@@ -34,6 +34,7 @@ class StudentDAO{
             }
         }
     }
+    
     static func getImage(_ userId: String, imageView: UIImageView){
            // Establish the connection with the database
            let db = Firestore.firestore()
@@ -61,20 +62,20 @@ class StudentDAO{
                    switch result{
                        
                    case .success(_):
-                       print("succesfully got image to the imageview")
+                       print("Succesfully got image to the imageview")
                    case .failure(_):
                        print("ERROR coudn't get image to the imageview")
                    }
                })
               } else {
                   
-                  print("did not find document for the url in getImage studentDAO")
+                  print("Did not find document for the url in getImage studentDAO")
               }
-           
-           }
-       }
-     static func setImage(_ userId: String, _ image: UIImage, completion: @escaping ()-> Void){
-         print("going to upload picture to the db")
+        }
+    }
+    
+    static func setImage(_ userId: String, _ image: UIImage, completion: @escaping ()-> Void){
+         print("Going to upload picture to the db")
          let data = image.jpegData(compressionQuality: 0.1)
          
          
@@ -113,14 +114,11 @@ class StudentDAO{
                          print("succesfully added image to the db")
                          completion()
                          
-                     }
-         }
-     }
-         
+                    }
+            }
+        }
+    }
     
-         
-         
-     }
     // Insert the student data in the database.
     // It is used a callback because we depend of the 'result' provided by the setData() function.
     static func addStudent(id: String, student: Student, completion: @escaping((_ data: String?) -> Void)){
@@ -191,11 +189,18 @@ class StudentDAO{
         // Establish the connection with the database
         let db = Firestore.firestore()
         
-        // Delete the student information in the database
-        db.collection("students").document("\(id)").delete()
-        
-        // Delete the information from authentication
-        Auth.auth().currentUser?.delete()
+        // Delete the student id from offers
+        JobOffersDAO.deleteInterestStudentId(id) { (error) in
+            if error != nil {
+                print("Error deleting student from all job offers")
+            } else {
+                // Delete the student information in the database
+                db.collection("students").document("\(id)").delete()
+                
+                // Delete the information from authentication
+                Auth.auth().currentUser?.delete()
+            }
+        }
     }
     
     // Get the user id
@@ -225,7 +230,6 @@ class StudentDAO{
         }
     }
        
-    
     // Get the general information of the student
     static func getStudent(_ userId: String, completion: @escaping(((String?), (Student?)) -> Void)) {
     
@@ -262,7 +266,6 @@ class StudentDAO{
             }
         }
     }
-    
     
     // Get a set of students given their student ids
     static func getStudents(_ studentIds: [String], completion: @escaping(([Student]?, [String]?) -> Void)) {
@@ -308,16 +311,16 @@ class StudentDAO{
                     // array to keep the same order in the students and in their ids
                     newStudentsIds.append(document.documentID)
                     
+                } else {
+                    db.collection("students").document(studentId).delete();
                 }
                 
                 if c == studentIds.count {
                     completion(students, newStudentsIds)
                 }
-                
             }
         }
     }
-    
     
     // Get a set of filtered students given their student ids and a speciality field
     static func getFilteredStudents(_ studentIds: [String], _ specialityField: String, completion: @escaping(([Student]?, [String]?) -> Void)) {
@@ -376,7 +379,6 @@ class StudentDAO{
             }
         }
     }
-    
     
     // Delete all the selected offer's notifications from a set of students caused by the deletion of an offer
     static func deleteNotifications(_ offerKey: String, _ interestedStudents: [String]){
